added a Readme.md to document the project goals
added a criteria folder and  goal PNGs get by
The Odin Project guidelines

created an index.html file for the landing page and
a style.css for the css style of the landing page
and any other pages that may be included.

the index contains the skeleton structure of the outward
div structure along with the header text
the style.css contains all the current properties along with
some improvised properties inc the font type.